requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile requirements.in
#
aiohappyeyeballs==2.6.1
    # via aiohttp
aiohttp==3.12.15
    # via
    #   langchain-azure-ai
    #   langchain-community
aiosignal==1.4.0
    # via aiohttp
altair==5.5.0
    # via streamlit
annotated-types==0.7.0
    # via pydantic
anyio==4.10.0
    # via
    #   httpx
    #   openai
async-timeout==4.0.3
    # via
    #   aiohttp
    #   langchain
attrs==25.3.0
    # via
    #   aiohttp
    #   jsonschema
    #   referencing
azure-ai-agents==1.1.0
    # via
    #   azure-ai-projects
    #   langchain-azure-ai
azure-ai-inference[opentelemetry]==1.0.0b9
    # via
    #   -r requirements.in
    #   langchain-azure-ai
azure-ai-projects==1.0.0b12
    # via langchain-azure-ai
azure-common==1.1.28
    # via azure-search-documents
azure-core==1.35.0
    # via
    #   azure-ai-agents
    #   azure-ai-inference
    #   azure-ai-projects
    #   azure-core-tracing-opentelemetry
    #   azure-cosmos
    #   azure-identity
    #   azure-search-documents
    #   azure-storage-blob
    #   langchain-azure-ai
azure-core-tracing-opentelemetry==1.0.0b12
    # via azure-ai-inference
azure-cosmos==4.14.0b2
    # via langchain-azure-ai
azure-identity==1.24.0
    # via langchain-azure-ai
azure-search-documents==11.5.3
    # via langchain-azure-ai
azure-storage-blob==12.26.0
    # via azure-ai-projects
beautifulsoup4==4.13.5
    # via bs4
blinker==1.9.0
    # via streamlit
bs4==0.0.2
    # via -r requirements.in
cachetools==6.2.0
    # via streamlit
certifi==2025.8.3
    # via
    #   httpcore
    #   httpx
    #   requests
cffi==1.17.1
    # via cryptography
charset-normalizer==3.4.3
    # via requests
click==8.2.1
    # via streamlit
cryptography==45.0.6
    # via
    #   azure-identity
    #   azure-storage-blob
    #   msal
    #   pyjwt
dataclasses-json==0.6.7
    # via langchain-community
distro==1.9.0
    # via openai
exceptiongroup==1.3.0
    # via
    #   anyio
    #   pytest
faiss-cpu==1.12.0
    # via -r requirements.in
frozenlist==1.7.0
    # via
    #   aiohttp
    #   aiosignal
gitdb==4.0.12
    # via gitpython
gitpython==3.1.45
    # via streamlit
greenlet==3.2.4
    # via sqlalchemy
h11==0.16.0
    # via httpcore
httpcore==1.0.9
    # via httpx
httpx==0.28.1
    # via
    #   langsmith
    #   openai
httpx-sse==0.4.1
    # via langchain-community
idna==3.10
    # via
    #   anyio
    #   httpx
    #   requests
    #   yarl
importlib-metadata==8.7.0
    # via opentelemetry-api
iniconfig==2.1.0
    # via pytest
isodate==0.7.2
    # via
    #   azure-ai-agents
    #   azure-ai-inference
    #   azure-ai-projects
    #   azure-search-documents
    #   azure-storage-blob
jinja2==3.1.6
    # via
    #   altair
    #   pydeck
jiter==0.10.0
    # via openai
jsonpatch==1.33
    # via langchain-core
jsonpointer==3.0.0
    # via jsonpatch
jsonschema==4.25.1
    # via altair
jsonschema-specifications==2025.4.1
    # via jsonschema
langchain==0.3.27
    # via
    #   -r requirements.in
    #   langchain-community
langchain-azure-ai==0.1.5
    # via -r requirements.in
langchain-community==0.3.29
    # via -r requirements.in
langchain-core==0.3.75
    # via
    #   -r requirements.in
    #   langchain
    #   langchain-azure-ai
    #   langchain-community
    #   langchain-openai
    #   langchain-text-splitters
langchain-openai==0.3.32
    # via
    #   -r requirements.in
    #   langchain-azure-ai
langchain-text-splitters==0.3.10
    # via langchain
langsmith==0.4.20
    # via
    #   langchain
    #   langchain-community
    #   langchain-core
lxml==6.0.1
    # via -r requirements.in
markupsafe==3.0.2
    # via jinja2
marshmallow==3.26.1
    # via dataclasses-json
msal==1.33.0
    # via
    #   azure-identity
    #   msal-extensions
msal-extensions==1.3.1
    # via azure-identity
multidict==6.6.4
    # via
    #   aiohttp
    #   yarl
mypy-extensions==1.1.0
    # via typing-inspect
narwhals==2.2.0
    # via altair
nest-asyncio==1.6.0
    # via -r requirements.in
numpy==2.2.6
    # via
    #   faiss-cpu
    #   langchain-azure-ai
    #   langchain-community
    #   pandas
    #   pydeck
    #   streamlit
openai==1.102.0
    # via
    #   -r requirements.in
    #   langchain-openai
opentelemetry-api==1.36.0
    # via azure-core-tracing-opentelemetry
orjson==3.11.3
    # via langsmith
packaging==25.0
    # via
    #   altair
    #   faiss-cpu
    #   langchain-core
    #   langsmith
    #   marshmallow
    #   pytest
    #   streamlit
pandas==2.3.2
    # via
    #   -r requirements.in
    #   streamlit
pillow==11.3.0
    # via streamlit
pluggy==1.6.0
    # via pytest
propcache==0.3.2
    # via
    #   aiohttp
    #   yarl
protobuf==6.32.0
    # via streamlit
pyarrow==21.0.0
    # via streamlit
pycparser==2.22
    # via cffi
pydantic==2.11.7
    # via
    #   langchain
    #   langchain-core
    #   langsmith
    #   openai
    #   pydantic-settings
pydantic-core==2.33.2
    # via pydantic
pydantic-settings==2.10.1
    # via langchain-community
pydeck==0.9.1
    # via streamlit
pygments==2.19.2
    # via pytest
pyjwt[crypto]==2.10.1
    # via
    #   msal
    #   pyjwt
pypdf==6.0.0
    # via -r requirements.in
pytest==8.4.1
    # via -r requirements.in
python-dateutil==2.9.0.post0
    # via pandas
python-dotenv==1.1.1
    # via
    #   -r requirements.in
    #   pydantic-settings
pytz==2025.2
    # via pandas
pyyaml==6.0.2
    # via
    #   langchain
    #   langchain-community
    #   langchain-core
referencing==0.36.2
    # via
    #   jsonschema
    #   jsonschema-specifications
regex==2025.7.34
    # via tiktoken
requests==2.32.5
    # via
    #   azure-core
    #   langchain
    #   langchain-community
    #   langsmith
    #   msal
    #   requests-toolbelt
    #   streamlit
    #   tiktoken
requests-toolbelt==1.0.0
    # via langsmith
rpds-py==0.27.1
    # via
    #   jsonschema
    #   referencing
six==1.17.0
    # via
    #   azure-core
    #   python-dateutil
smmap==5.0.2
    # via gitdb
sniffio==1.3.1
    # via
    #   anyio
    #   openai
soupsieve==2.8
    # via beautifulsoup4
sqlalchemy==2.0.43
    # via
    #   langchain
    #   langchain-community
streamlit==1.49.0
    # via -r requirements.in
tenacity==9.1.2
    # via
    #   langchain-community
    #   langchain-core
    #   streamlit
tiktoken==0.11.0
    # via
    #   -r requirements.in
    #   langchain-openai
toml==0.10.2
    # via streamlit
tomli==2.2.1
    # via pytest
tornado==6.5.2
    # via streamlit
tqdm==4.67.1
    # via openai
typing-extensions==4.15.0
    # via
    #   aiosignal
    #   altair
    #   anyio
    #   azure-ai-agents
    #   azure-ai-inference
    #   azure-ai-projects
    #   azure-core
    #   azure-cosmos
    #   azure-identity
    #   azure-search-documents
    #   azure-storage-blob
    #   beautifulsoup4
    #   exceptiongroup
    #   langchain-core
    #   multidict
    #   openai
    #   opentelemetry-api
    #   pydantic
    #   pydantic-core
    #   pypdf
    #   referencing
    #   sqlalchemy
    #   streamlit
    #   typing-inspect
    #   typing-inspection
typing-inspect==0.9.0
    # via dataclasses-json
typing-inspection==0.4.1
    # via
    #   pydantic
    #   pydantic-settings
tzdata==2025.2
    # via pandas
urllib3==2.5.0
    # via requests
watchdog==6.0.0
    # via streamlit
yarl==1.20.1
    # via aiohttp
zipp==3.23.0
    # via importlib-metadata
zstandard==0.24.0
    # via langsmith

# The following packages are considered to be unsafe in a requirements file:
# pip
